Automated metadata reader
Uses QTKit bridge to read metadata from QT files. File names and
metadata are stored in an XLS sheet for reference. Metadata from files
processed during batch is exported to a text file to be read using a VB
script. The VB script crams the data into the MVIS system.